refactor: introduce Rust-only Engine wrapper over extract internals
Add a `crate::engine` module with `Engine`/`EngineBuilder` and move the
extraction internals (extract_one, extract_remote_uri, batch concurrency,
scrape/crawl output converters) out of `core/extract/mod.rs` into it. The
free `extract`/`extract_batch` functions become thin delegators over a
process-global default Engine — signatures and behavior are unchanged.

The engine module is Rust-only: it is a bare `pub mod` whose files are not
listed in alef.toml `sources`, so the binding generator emits nothing for
it, and Engine/EngineBuilder are also added to the alef exclude list as a
safeguard. This is a pure refactor; the single-URL path still builds a
fresh CrawlEngine (batch reuse comes later).
